SQL improvements
