+++ app.py
@@ -6,7 +6,7 @@ app = Flask(__name__)
 
 @app.route('/', methods=['GET','POST'])
 def login():
-    # Main flask interface
+    # Main flask map interface
     pass
 
 @app.route('/api/', methods=['GET', 'POST'])
@@ -17,7 +17,7 @@ def add_message(token=None):
     except:
         print("Format data not valid.")
 
-    # Get authorisation
+    # Verify identity of the uploader
     with open("users.json", "r") as f:
         auth = json.load(f)
     user = request.args.get("user")
@@ -29,7 +29,7 @@ def add_message(token=None):
     else:
         access=False
 
-    # Verify identity of the uploader
+    # Main script
     if access:
         data_processor(user, content)
         return jsonify({"result": "ok"})
