+++ data_processor.py
@@ -1,7 +1,8 @@
 import pandas as pd
 from pathlib import Path
-import gps_utils
 import json
+import sql_connection
+import sql_dump
 
 
 def flatten_json(nested_json, exclude=['']):
@@ -30,45 +31,9 @@ def flatten_json(nested_json, exclude=['']):
     return out
 
 
-def get_user_dir(user_id):
-    dir = Path.cwd()
-    return dir/'storage'/user_id
-
-
-def directories(user_id):
-    # Get or create user files if they don't exist
-
-    # User directory
-    user_dir = get_user_dir(user_id)
-    user_dir.mkdir(parents=True, exist_ok=True)
-
-    # User's csv file to store all GPS points
-    file = user_id+'.csv'
-    filepath = user_dir / file
-    if not filepath.is_file():
-        df = pd.DataFrame(columns=['type','geometry_type','geometry_coordinates_0','geometry_coordinates_1','properties_speed','properties_battery_state','properties_timestamp','properties_battery_level','properties_vertical_accuracy','properties_pauses','properties_horizontal_accuracy','properties_wifi','properties_deferred','properties_significant_change','properties_locations_in_payload','properties_activity','properties_device_id','properties_altitude','properties_desired_accuracy','properties_motion_0','properties_action','properties_motion_1','properties_arrival_date','properties_departure_date'])
-        df.to_csv(filepath, index=False)
-    
-    # User's reduced csv file
-    file = user_id+'_min.csv'
-    filepath = user_dir / file
-    if not filepath.is_file():
-        df = pd.DataFrame(columns=['LONG','LAT','ALT','speed','timestamp','vertical_accuracy','horizontal_accuracy','motion'])
-        df.to_csv(filepath, index=False)
-
-    # User's visits csv
-    file =  user_id+'_visits.csv'
-    filepath = user_dir / file
-    if not filepath.is_file():
-        df = pd.DataFrame(columns=['LONG','LAT','arrival','departure','device'])
-        df.to_csv(filepath, index=False)
-    
-    return True
-
-
 def write_json(user_id, content):
     file = user_id+'_last.json'
-    filepath = get_user_dir(user_id) / file
+    filepath = Path.cwd()/'storage'/file
     with open(filepath,'w') as file:
         json.dump(content, file, indent = 4)
 
@@ -113,9 +78,10 @@ def get_df(user_id, content):
 
 def data_processor(user_id, content):
     
-    directories(user_id)  # Creates directories and files if it's new user
-    write_json(user_id, content)  # Saves a copy of the received content
+    write_json(user_id, content)  # Saves a copy of the last received content
     df = get_df(user_id, content)  # Gets df from json content
+
+    sql_connection.main(user_id)  # Creates a new sql connection
 
     # Gather other GPS data
     gps_utils.checkins(df, user_id)
@@ -127,4 +93,4 @@ def data_processor(user_id, content):
 
 
 if __name__ == "__main__":
-    print(directories('daniel'))
+    pass
